:sparkles: Nav: Abstracting to hook
